Move router to its own separate crate
