+++ crates/http/src/lib.rs
@@ -3,12 +3,12 @@ pub use wasmtime_wasi_http::body::HyperIncomingBody as Body;
 
 pub mod app_info;
 pub mod config;
-pub mod routes;
 pub mod trigger;
 #[cfg(feature = "runtime")]
 pub mod wagi;
 
-pub const WELL_KNOWN_PREFIX: &str = "/.well-known/spin/";
+pub use spin_http_routes as routes;
+pub use spin_http_routes::WELL_KNOWN_PREFIX;
 
 #[cfg(feature = "runtime")]
 pub mod body {
